Three unsigned fields the retail loads prove: the DIM explosion ray angles, the planet-select lock and the dll_200 hug counter
Retail load census per struct offset:

  DimExplosionState 0xA44/0xA46 rayYawA, rayPitchA  4 loads, all lhz, no
    lha anywhere; the +0x4000 wrap is followed by clrlwi r0,r0,16, and the
    only sign extension is the explicit (s16) at the anim.rotX/rotY store
  WorldPlanetState 0x09 selectionLocked  5 loads, 2 read straight into a
    cmplwi, none sign-extended
  Dll200State 0x27 counter27             3 loads, one cmplwi r0,2, the
    other two incremented straight back into a stb

The B halves of the ray pair are stored only, and carry the type of the A
halves they are copied from. Retype the six fields unsigned and drop the
nine reader casts and address puns that were carrying it by hand.

Every object byte-identical; main.dol OK.

+++ src/dlls/objects/466_WORLDplanet/WORLDplanet.c
@@ -341,7 +341,7 @@ void worldplanet_update(GameObject* obj) {
             }
             state->unlockedPlanetMask = z[1];
         }
-        if (gWorldPlanetSelectConfirmTimer == 0 && (u8)state->selectionLocked == 0) {
+        if (gWorldPlanetSelectConfirmTimer == 0 && state->selectionLocked == 0) {
             while (!done) {
                 state->selectedPlanet = state->selectedPlanet + in.inX[0];
                 if (state->selectedPlanet < 0) {
@@ -392,7 +392,7 @@ void worldplanet_update(GameObject* obj) {
             pstate = planet->extra;
             planet->anim.rotY = (obj)->anim.rotY;
             planet->anim.rotX = (obj)->anim.rotX;
-            if ((u8)state->selectionLocked != 0 || (((int)(u32)state->unlockedPlanetMask >> i) & 1) == 0) {
+            if (state->selectionLocked != 0 || (((int)(u32)state->unlockedPlanetMask >> i) & 1) == 0) {
                 pstate->effectState = 0;
                 if ((int)i == state->selectedPlanet) {
                     arwing->anim.flags |= OBJANIM_FLAG_HIDDEN;
@@ -439,7 +439,7 @@ void worldplanet_update(GameObject* obj) {
         }
         objId = (int)ObjList_FindObjectById(tbl->orbitObjectIds[gWorldPlanetSelectionToIndex[state->selectedPlanet]]);
         if (getLoadedFileFlags(WORLDPLANET_SAVE_FILE_SLOT) == 0 && gWorldPlanetInputLockTimer == 0) {
-            switch ((u8)state->selectionLocked) {
+            switch (state->selectionLocked) {
             case 0:
                 if (gWorldPlanetReselectDelayTimer != 0) {
                     gWorldPlanetReselectDelayTimer -= 1;
